Move test case to test_builtin
